backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
python-dotenv==1.0.1
httpx==0.27.2
python-multipart==0.0.9
redis==5.0.1
aioredis==2.0.1
PyJWT==2.8.0
cryptography==42.0.8
websockets==12.0
fastapi-websocket==0.0.3
passlib[bcrypt]==1.7.4
numpy==1.24.3
scikit-learn==1.3.0
pytest==8.3.4
pytest-cov==4.1.0
